Fix function passthrough
